rework: Refactored getting daily emissions
Extracted Daily Emissions function into its own separate class.
