+++ app/src/main/java/com/example/b07projectfall2024/NavigationBar/DailyEmissionGetterReceiver.java
@@ -1,0 +1,10 @@
+package com.example.b07projectfall2024.NavigationBar;
+
+public interface DailyEmissionGetterReceiver {
+
+    public void TransportAction(double AdditionalEmissions, String Date);
+    public void FoodAction(double AdditionalEmissions, String Date);
+    public void ConsumptionAction(double AdditionalEmissions, String Date);
+    public void init();
+
+}
